// src/lib-spare/models/PurchaseOrderLineDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PurchaseOrderLineDto = {
    id?: string;
    productId?: string;
    productName?: string;
    sku?: string;
    qty?: number;
    unitPrice?: number;
};