// src/lib-spare/models/ReceiptLineDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ReceiptLineDto = {
    id?: string;
    poLineId?: string;
    productId?: string;
    productName?: string;
    sku?: string;
    orderedQty?: number;
    receivedQty?: number;
    note?: string;
};