// src/lib-spare/models/ProductLocationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductLocationDto = {
    agencyId?: string;
    productId?: string;
    binCode: string;
    quantity?: number;
    note?: string;
    updatedAt?: string;
};